Optimized ascend test framework for e2e test.
