+++ python/sglang/test/ascend/e2e/k8s_multi_pd_mix.yaml.jinja2
@@ -1,3 +1,11 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ kube_config_map }}
+  namespace: {{ name_space }}
+data: {}
+
+---
 apiVersion: batch.volcano.sh/v1alpha1
 kind: Job
 metadata:
@@ -7,20 +15,21 @@ metadata:
     ring-controller.atlas: ascend-1980
     fault-scheduling: "force"
 spec:
-  minAvailable: 1
+  minAvailable: {{ node_size }}
   schedulerName: volcano
   policies:
     - event: PodEvicted
       action: RestartJob
   tasks:
-  - name: "pod"
-    replicas: 1
+  - name: "sglang-node"
+    replicas: {{ node_size }}
     template:
       metadata:
         labels:
           app: sgl-ascend
           ring-controller.atlas: ascend-1980
       spec:
+        hostNetwork: True
         containers:
         - image: {{ image }}
           imagePullPolicy: Always
@@ -43,40 +52,35 @@ spec:
           - name: KUBECONFIG
             value: "{{ kube_config }}"
           - name: NAMESPACE
-            value: "ascend-sglang-single"
+            value: "{{ name_space }}"
+          - name: KUBE_CONFIG_MAP
+            value: "{{ kube_config_map }}"
           - name: HF_ENDPOINT
             value: "https://hf-mirror.com"
           - name: SGLANG_IS_IN_CI
-            value: : "true"
+            value: "{{ sglang_is_in_ci }}"
           command: ["/bin/bash", "-c"]
           args:
             - |
-               bash /root/sglang/python/sglang/test/ascend/performance/run_ascend_testcase.sh {{ test_case }}
+               bash /root/sglang/python/sglang/test/ascend/e2e/run_ascend_testcase.sh {{ test_case }}
           resources:
             requests:
-              huawei.com/ascend-1980: {{ npu_num }}
-              cpu: "76"
+              huawei.com/ascend-1980: 16
             limits:
-              huawei.com/ascend-1980: {{ npu_num }}
-              cpu: "76"
+              huawei.com/ascend-1980: 16
           volumeMounts:
           - name: ascend-driver
             mountPath: /usr/local/Ascend/driver
           - name: localtime
             mountPath: /etc/localtime
-          - name: kube-config
-            mountPath: /root/script
           - name: share
-            mountPath: /data
+            mountPath: /data/ascend-ci-share-pkking-sglang
           - name: share
             mountPath: /root/.cache
           - name: share
             mountPath: /root/sglang
-            subPath: {{ sglang_source_path }}
+            subPath: {{ sglang_source_relative_path }}
         volumes:
-        - name: kube-config
-          hostPath:
-            path: /root/script
         - name: share
           persistentVolumeClaim:
             claimName: sglang-guiyang004
@@ -86,6 +90,17 @@ spec:
         - name: localtime
           hostPath:
             path: /etc/localtime
+        {% if env == "debug" %}
+        tolerations:
+          - key: "instance"
+            operator: "Equal"
+            value: "npu-class-service"
+            effect: "NoSchedule"
+        {% endif %}
         nodeSelector:
           accelerator/huawei-npu: ascend-snt9c
+          {% if env == "debug" %}
+          nodestatus: debug
+          {% endif %}
         restartPolicy: OnFailure
+
