--- python/sglang/test/ascend/performance/k8s_multi_pd_mix.yaml.jinja2
@@ -1,173 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: rings-config-ascend-sglang-test
-  namespace: {{ name_space }}
-  labels:
-    ring-controller.atlas: ascend-1980
-data:
-  hccl.json: |
-    {
-        "status":"initializing"
-    }
-
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: {{ kube_config_map }}
-  namespace: {{ name_space }}
-data: {}
-
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: Role
-metadata:
-  name: ascend-sglang-configmap-access
-  namespace: {{ name_space }}
-rules:
-- apiGroups: [""]
-  resources: ["configmaps"]
-  resourceNames: ["{{ kube_config_map }}"]
-  verbs: ["get", "update", "patch"]
-
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: ascend-sglang-configmap-binding
-  namespace: {{ name_space }}
-subjects:
-- kind: ServiceAccount
-  name: default
-  namespace: {{ name_space }}
-roleRef:
-  kind: Role
-  name: ascend-sglang-configmap-access
-  apiGroup: rbac.authorization.k8s.io
-
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: Role
-metadata:
-  name: ascend-sglang-pod-access
-  namespace: {{ name_space }}
-rules:
-- apiGroups: [""]
-  resources: ["pods"]
-  verbs: ["get", "list"]
-
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: ascend-sglang-pod-binding
-  namespace: {{ name_space }}
-subjects:
-- kind: ServiceAccount
-  name: default
-  namespace: {{ name_space }}
-roleRef:
-  kind: Role
-  name: ascend-sglang-pod-access
-  apiGroup: rbac.authorization.k8s.io
-
----
-apiVersion: batch.volcano.sh/v1alpha1
-kind: Job
-metadata:
-  name: {{ kube_job_name }}
-  namespace: {{ name_space }}
-  labels:
-    ring-controller.atlas: ascend-1980
-    fault-scheduling: "force"
-spec:
-  minAvailable: {{ node_size }}
-  schedulerName: volcano
-  policies:
-    - event: PodEvicted
-      action: RestartJob
-  tasks:
-  - name: "sglang-node"
-    replicas: {{ node_size }}
-    template:
-      metadata:
-        labels:
-          app: sgl-ascend
-          ring-controller.atlas: ascend-1980
-      spec:
-        hostNetwork: True
-        containers:
-        - image: {{ image }}
-          imagePullPolicy: Always
-          securityContext:
-            privileged: true
-          name: sgl-ascend
-          env:
-          - name: HOSTNAME
-            valueFrom:
-              fieldRef:
-                fieldPath: metadata.name
-          - name: POD_IP
-            valueFrom:
-              fieldRef:
-                fieldPath: status.hostIP
-          - name: INSTALL_SGLANG_FROM_SOURCE
-            value: "{{ install_sglang_from_source }}"
-          - name: METRICS_DATA_FILE
-            value: "{{ metrics_data_file }}"
-          - name: KUBECONFIG
-            value: "{{ kube_config }}"
-          - name: NAMESPACE
-            value: "{{ name_space }}"
-          - name: KUBE_CONFIG_MAP
-            value: "{{ kube_config_map }}"
-          - name: HF_ENDPOINT
-            value: "https://hf-mirror.com"
-          - name: SGLANG_IS_IN_CI
-            value: "{{ sglang_is_in_ci }}"
-          command: ["/bin/bash", "-c"]
-          args:
-            - |
-               bash /root/sglang/python/sglang/test/ascend/performance/run_ascend_testcase.sh {{ test_case }}
-          resources:
-            requests:
-              huawei.com/ascend-1980: 16
-            limits:
-              huawei.com/ascend-1980: 16
-          volumeMounts:
-          - name: ascend-910-config
-            mountPath: /user/serverid/devindex/config
-          - name: ascend-driver
-            mountPath: /usr/local/Ascend/driver
-          - name: localtime
-            mountPath: /etc/localtime
-          - name: kube-config
-            mountPath: /root/script
-          - name: share
-            mountPath: /data
-          - name: share
-            mountPath: /root/.cache
-          - name: share
-            mountPath: /root/sglang
-            subPath: {{ sglang_source_path }}
-        volumes:
-        - name: kube-config
-          hostPath:
-            path: /root/script
-        - name: share
-          persistentVolumeClaim:
-            claimName: sglang-guiyang004
-        - name: ascend-910-config
-          configMap:
-            name: rings-config-ascend-sglang-test
-        - name: ascend-driver
-          hostPath:
-            path: /usr/local/Ascend/driver
-        - name: localtime
-          hostPath:
-            path: /etc/localtime
-        nodeSelector:
-          accelerator/huawei-npu: ascend-snt9c
-        restartPolicy: OnFailure
-
