+++ test/manual/ascend/performance/test_ascend_deepseek_r1_w8a8_2p1d_32p_in6k_out1k6_15ms.py
@@ -1,10 +1,11 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestAscendMultiNodePdSepTestCaseBase,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerfMultiNodePdSepTestCaseBase,
     DEEPSEEK_R1_W8A8_MODEL_PATH,
-    NIC_NAME, ROUND_ROBIN
+    ROUND_ROBIN
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 
 MODEL_CONFIG = {
     "model_path": DEEPSEEK_R1_W8A8_MODEL_PATH,
@@ -91,7 +92,7 @@ MODEL_CONFIG = {
     ],
 }
 
-class TestDeepSeekR1W8A8(TestAscendMultiNodePdSepTestCaseBase):
+class TestDeepSeekR1W8A8(TestAscendPerfMultiNodePdSepTestCaseBase):
     model_config = MODEL_CONFIG
     dataset_name = "random"
     request_rate = 16
