+++ test/manual/ascend/performance/test_ascend_deepseek_v3_2_w8a8_1p1d_32p_in64k_out3k_30ms.py
@@ -1,10 +1,11 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestAscendMultiNodePdSepTestCaseBase,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerfMultiNodePdSepTestCaseBase,
     DEEPSEEK_V32_W8A8_MODEL_PATH,
-    NIC_NAME, ROUND_ROBIN
+    ROUND_ROBIN
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 
 MODEL_CONFIG = {
     "model_path": DEEPSEEK_V32_W8A8_MODEL_PATH,
@@ -93,7 +94,7 @@ MODEL_CONFIG = {
     ],
 }
 
-class TestDeepSeekV32(TestAscendMultiNodePdSepTestCaseBase):
+class TestDeepSeekV32(TestAscendPerfMultiNodePdSepTestCaseBase):
     model_config = MODEL_CONFIG
     dataset_name = "random"
     request_rate = "inf"
