+++ test/manual/ascend/performance/test_ascend_qwen3_235b_w8a8_16p_in2k_out2k_50ms.py
@@ -1,11 +1,11 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestMultiNodePdMixTestCaseBase,
-    NIC_NAME,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerfMultiNodePdMixTestCaseBase,
     QWEN3_235B_A22B_EAGLE_MODEL_PATH,
     QWEN3_235B_W8A8_MODEL_PATH
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 
 MODEL_CONFIG = {
     "model_path": QWEN3_235B_W8A8_MODEL_PATH,
@@ -49,7 +49,7 @@ MODEL_CONFIG = {
     ]
 }
 
-class TestQwen235B(TestMultiNodePdMixTestCaseBase):
+class TestQwen235B(TestAscendPerfMultiNodePdMixTestCaseBase):
     model_config = MODEL_CONFIG
     dataset_name = "random"
     max_concurrency = 768
