+++ test/manual/ascend/performance/test_ascend_qwen3_235b_w8a8_1p1d_24p_in3k5_out1k5_50ms.py
@@ -1,11 +1,11 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestAscendMultiNodePdSepTestCaseBase,
-    NIC_NAME,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerfMultiNodePdSepTestCaseBase,
     QWEN3_235B_A22B_EAGLE_MODEL_PATH,
     QWEN3_235B_W8A8_MODEL_PATH
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 
 MODEL_CONFIG = {
     "model_path": QWEN3_235B_W8A8_MODEL_PATH,
@@ -101,7 +101,7 @@ MODEL_CONFIG = {
 }
 
 
-class TestQwen235bW8A8(TestAscendMultiNodePdSepTestCaseBase):
+class TestQwen235bW8A8(TestAscendPerfMultiNodePdSepTestCaseBase):
     model_config = MODEL_CONFIG
     backend = "sglang-oai"
     dataset_name = "random"
