+++ test/registered/ascend/performance/test_ascend_deepseek_r1_w4a8_8p_in3k5_out1k5_50ms.py
@@ -1,10 +1,10 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestPerformanceTestCaseBase,
-    NIC_NAME,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerformanceTestCaseBase,
     DEEPSEEK_R1_W4A8_PER_CHANNEL_MODEL_PATH
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(est_time=1800, suite="nightly-16-npu-a3", nightly=True)
@@ -54,7 +54,7 @@ MODEL_OTHER_ARGS = (
 )
 
 
-class TestAscendDeepSeekR1W4A8(TestPerformanceTestCaseBase):
+class TestAscendDeepSeekR1W4A8(TestAscendPerformanceTestCaseBase):
     model = DEEPSEEK_R1_W4A8_PER_CHANNEL_MODEL_PATH
     other_args = MODEL_OTHER_ARGS
     envs = MODEL_ENVS
