+++ test/registered/ascend/performance/test_ascend_qwen3_235b_w8a8_8p_in3k5_out1k5_50ms.py
@@ -1,11 +1,11 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestPerformanceTestCaseBase,
-    NIC_NAME,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerformanceTestCaseBase,
     QWEN3_235B_A22B_EAGLE_MODEL_PATH,
     QWEN3_235B_W8A8_MODEL_PATH
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(est_time=1800, suite="nightly-16-npu-a3", nightly=True)
@@ -53,7 +53,7 @@ QWEN3_235B_OTHER_ARGS = (
     ]
 )
 
-class TestQwen235B(TestPerformanceTestCaseBase):
+class TestQwen235B(TestAscendPerformanceTestCaseBase):
     model = QWEN3_235B_W8A8_MODEL_PATH
     other_args = QWEN3_235B_OTHER_ARGS
     envs = QWEN3_235B_ENVS
