+++ test/registered/ascend/performance/test_ascend_qwen3_30b_w8a8_1p_in3k5_out1k5_50ms.py
@@ -1,11 +1,11 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestPerformanceTestCaseBase,
-    NIC_NAME,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerformanceTestCaseBase,
     QWEN3_30B_A3B_W8A8_MODEL_PATH,
     QWEN3_A3B_EAGLE_MODEL_PATH
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(est_time=1800, suite="nightly-2-npu-a3", nightly=True)
@@ -46,7 +46,7 @@ QWEN3_30B_A3B_OTHER_ARGS = (
     ]
 )
 
-class TestQwen30B(TestPerformanceTestCaseBase):
+class TestQwen30B(TestAscendPerformanceTestCaseBase):
     model = QWEN3_30B_A3B_W8A8_MODEL_PATH
     other_args = QWEN3_30B_A3B_OTHER_ARGS
     envs = QWEN3_30B_A3B_ENVS
