+++ test/registered/ascend/performance/test_ascend_qwen3_480b_w8a8_8p_in3k5_out1k5_50ms.py
@@ -1,10 +1,10 @@
 import unittest
 
-from sglang.test.ascend.performance.test_ascend_performance_utils import (
-    TestPerformanceTestCaseBase,
+from sglang.test.ascend.e2e.test_ascend_performance_utils import (
+    TestAscendPerformanceTestCaseBase,
     QWEN3_480B_W8A8_MODEL_PATH,
-    NIC_NAME
 )
+from sglang.test.ascend.e2e.test_ascend_multi_node_utils import NIC_NAME
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(est_time=1800, suite="nightly-16-npu-a3", nightly=True)
@@ -41,7 +41,7 @@ QWEN3_480B_OTHER_ARGS = [
     "--cuda-graph-bs", 16, 20, 24,
 ]
 
-class TestQwen480B(TestPerformanceTestCaseBase):
+class TestQwen480B(TestAscendPerformanceTestCaseBase):
     model = QWEN3_480B_W8A8_MODEL_PATH
     other_args = QWEN3_480B_OTHER_ARGS
     envs = QWEN3_480B_ENVS
